backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
google-genai==0.7.0
PyMuPDF==1.22.5
pydantic==1.10.11